Refactor DatasetsContent to import DatasetsCenterContent from the correct path for improved module organization
